requirements.txt
requests
Pillow
numpy
torch